refactor: rename /lib to /src
