+++ package.json
@@ -25,9 +25,9 @@
     "webpack": "nrfconnect-scripts build-dev",
     "build": "nrfconnect-scripts build-prod",
     "nordic-publish": "nrfconnect-scripts nordic-publish",
-    "lint": "nrfconnect-scripts lint lib index.jsx",
+    "lint": "nrfconnect-scripts lint src index.jsx",
     "lint-init": "nrfconnect-scripts lint-init",
-    "lintfix": "nrfconnect-scripts lint --fix lib index.jsx",
+    "lintfix": "nrfconnect-scripts lint --fix src index.jsx",
     "test": "nrfconnect-scripts test",
     "test-watch": "nrfconnect-scripts test --watch",
     "clean": "npm run clean-dist && npm run clean-modules",
