+++ src/actions/deviceActions.js
