+++ src/actions/settingsActions.js
