+++ src/components/AppSidePanelView.jsx
