+++ src/components/ChannelView.jsx
