+++ src/components/ChartView.jsx
