+++ src/components/OtherSettingsView.jsx
