+++ src/components/PacketView.jsx
