+++ src/components/ToggleChannelModeView.jsx
