+++ src/containers/appSidePanelView.js
