+++ src/containers/channelView.js
