+++ src/containers/otherSettingsView.js
