+++ src/containers/toggleChannelModeView.js
