+++ src/reducers/deviceReducer.js
