+++ src/reducers/settingsReducer.js
